[FEAT] - Criação das Migrations

+++ DevFreela.Application/InputModels/UpdateProjectInputModel.cs
@@ -9,6 +9,6 @@ namespace DevFreela.Application.InputModels
         public int Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public decimal TotalCost { get; set; }
+        public double TotalCost { get; set; }
     }
 }
