+++ DevFreela.Core/Entities/Project.cs
@@ -7,7 +7,7 @@ namespace DevFreela.Core.Entities
 {
     public class Project : BaseEntity
     {
-        public Project(string title, string description, int idClient, int idFreelancer, decimal totalCost)
+        public Project(string title, string description, int idClient, int idFreelancer, double totalCost)
         {
             Title = title;
             Description = description;
@@ -26,7 +26,7 @@ namespace DevFreela.Core.Entities
         public User Client { get; private set; }
         public int IdFreelancer { get; private set; }
         public User Freelancer { get; private set; }
-        public decimal TotalCost { get; private set; }
+        public double TotalCost { get; private set; }
         public DateTime CreatedAt { get; private set; }
         public DateTime? StartedAt { get; private set; }
         public DateTime? FinishedAt { get; private set; }
@@ -59,7 +59,7 @@ namespace DevFreela.Core.Entities
             }
         }
 
-        public void Update(string title, string description, decimal totalCost)
+        public void Update(string title, string description, double totalCost)
         {
             Title = title;
             Description = description;
